+++ CHANGES.rst
@@ -1,6 +1,15 @@
 CHANGE LOG
 ==========
 
+2.0.1
+-----
+- Fixed bug forcing users to add mobile number when not required
+
+2.1.0
+-----
+- Update template to only show mobile number field if it's activated and has country calling code
+- Show a warning message on CMS if mobile number is activated but country calling code has not been set
+
 2.0.2
 -----
 - Added missing button class for templates